* Further fixes for #16 - the levels of indirection on the max speed slider were absurd.  Now the slider goes in 1 step/second increments!
* reverting c4ba8d40b4317325861e93286cbdfceda67283d3

+++ Graffik/gui/film/JogControlManager/jogcontrolpanel.h
@@ -87,7 +87,7 @@ private slots:
     void _motorStarted(unsigned short p_addr);
     void _motorStopped(unsigned short p_addr);
     void _jogDampChanged(int p_val);
-    void _jogSpeedChanged(double p_val);
+    void _jogSpeedChanged(int p_val);
 
 private:
     Ui::JogControlPanel *ui;
